V0.04 Quest In Progress!
Quest are here but not here see
the backend of the code is there and is working BUT
no one in the game is giving out a quest right now...
But will soon! [one is going to be to get all three parts to the
teleporter and "fix" it more to come about that later]
Anyway the gist is the quest are 50% of the way here and will be in the
game soon!