V0.045 Quest Book or at lest some of it.
so note how we got a extra 5 in this patch?
Thats becuse this was a little harder than expected i added some user
end (you can now see the UI talk to the back end and the end output) and
i also added NEW backend aka (buttons talking to the backend that now
can hear the buttons) LONG story anyway this is good becuse i added some
parts to the scripts so that they could give some info about quest your
on this may mean that down the road NPCs will use this data and
formulate some dilog based on this data. anyway to end this update i
know its not what i wanted to push but its all i got for now later i
will add more Quest Ui in tell then have fun with the parts that are in
there.